🐛 Don't write default JSON values via pydantic
There was a bug with the Haskell OpenAPI generator. Specifically:

Error in $.chemicals[0].attributi[0]['attributo_value_bool']:
  expected Bool, but encountered Null

Which was correct, the attributo wasn't of type bool, so the
value_bool field was serialized by pydantic as null (instead of
omitting it).

I added the necessary setting to every endpoint now, resulting in a
decrease in data for read_runs of 1.3MiB → 300KiB, so quite a lot.
Probably this means more performance in the client as well.

The change doesn't seem to affect the frontend or the generated
openapi.json.